+++ out/subscriptions/TyranoCompletionItemProvider.js
@@ -4,6 +4,7 @@ exports.TyranoCompletionItemProvider = void 0;
 const fs = require("fs");
 const vscode = require("vscode");
 const InformationWorkSpace_1 = require("../InformationWorkSpace");
+const path = require("path");
 /**
  * [1]プロジェクト中に存在する素材（画像、音声、シナリオ、外部JSを読み込みスニペット登録
  * →おそらくワークスペースに変更が加わるたびにワークスペースに更新掛ける必要がある。
@@ -15,9 +16,10 @@ const InformationWorkSpace_1 = require("../InformationWorkSpace");
  * 実装順は312
  */
 class TyranoCompletionItemProvider {
+    infoWs = InformationWorkSpace_1.InformationWorkSpace.getInstance();
+    tyranoTagSnippets; //公式タグのスニペット定義
+    info = InformationWorkSpace_1.InformationWorkSpace.getInstance();
     constructor() {
-        this.infoWs = InformationWorkSpace_1.InformationWorkSpace.getInstance();
-        this.info = InformationWorkSpace_1.InformationWorkSpace.getInstance();
         //タグスニペットファイル読み込み
         this.tyranoTagSnippets = JSON.parse(fs.readFileSync(__dirname + "/./../../snippet/tyrano.snippet.json", "utf8"));
     }
@@ -30,8 +32,11 @@ class TyranoCompletionItemProvider {
      * @returns
      */
     async provideCompletionItems(document, position, token, context) {
+        let projectPath = await this.infoWs.getProjectPathByFilePath(document.fileName);
+        let cursor = vscode.window.activeTextEditor?.selection.active.character;
         //カーソル付近のタグデータを取得
-        const parsedData = this.infoWs.parser.tyranoParser.parseScenario(document.lineAt(position.line).text);
+        const lineText = document.lineAt(position.line).text;
+        const parsedData = this.infoWs.parser.tyranoParser.parseScenario(lineText);
         const array_s = parsedData["array_s"];
         let tagNumber = "";
         for (let data in array_s) {
@@ -41,14 +46,26 @@ class TyranoCompletionItemProvider {
             }
             tagNumber = data;
         }
-        //空行orテキストならタグの予測変換を出す
-        if (array_s === undefined || array_s[tagNumber] === undefined) {
+        let tagParams = await vscode.workspace.getConfiguration().get('TyranoScript syntax.tag.parameter');
+        const leftSideText = array_s[tagNumber] !== undefined ? lineText.substring(array_s[tagNumber]["column"], cursor) : undefined;
+        const lineTagName = array_s[tagNumber] !== undefined ? array_s[tagNumber]["name"] : undefined; //今見てるタグの名前
+        const regExp2 = new RegExp('(\\S)+="(?![\\s\\S]*")', "g"); //今見てるタグの値を取得
+        let regExpResult = leftSideText?.match(regExp2); //「hoge="」を取得できる
+        let lineParamName = undefined;
+        if (regExpResult) {
+            lineParamName = regExpResult[0].replace("\"", "").replace("=", "").trim(); //今見てるパラメータの名前
+        }
+        const paramInfo = lineTagName !== undefined && tagParams[lineTagName] !== undefined ? tagParams[lineTagName][lineParamName] : undefined; //今見てるタグのパラメータ情報  paramsInfo.path paramsInfo.type
+        if (array_s[tagNumber] !== undefined && lineTagName !== undefined && lineParamName !== undefined && paramInfo !== undefined) {
+            return await this.completionResource(projectPath, paramInfo.type, projectPath + this.infoWs.pathDelimiter + paramInfo.path);
+        }
+        else if (array_s === undefined || array_s[tagNumber] === undefined) { //空行orテキストならタグの予測変換を出す
             return this.completionTag();
         }
         else { //タグの中ならタグのパラメータの予測変換を出す 
-            let isTagSentence = array_s[tagNumber]["name"] === "text" || array_s[tagNumber]["name"] === undefined ? false : true;
+            let isTagSentence = lineTagName === "text" || lineTagName === undefined ? false : true;
             if (isTagSentence) {
-                return this.completionParameter(array_s[tagNumber]["name"], array_s[tagNumber]["pm"]);
+                return this.completionParameter(lineTagName, array_s[tagNumber]["pm"]);
             }
             else {
                 return this.completionTag();
@@ -59,28 +76,52 @@ class TyranoCompletionItemProvider {
      * //TODO:ラベルの予測変換
      * *から始まる単語なら予測変換をだす
      */
-    conpletionLabel() {
+    completionLabel() {
         // comp.kind = vscode.CompletionItemKind.Variable;
+        // vscode.CompletionItemKind.Module
         //「target="」「target_cancel="」が直前にあるならラベルの一覧を候補表示
     }
     /**
      * //TODO:変数の予測変換
-     * f. sf. tf. のいずれかから始まった時予測変換を出す。
+     * [\s*(f.|sf.|tf.|mp.)]のいずれかから始まった時予測変換を出す。
      * InformationWorkSpaceに登録済みの変数リストを取得すれば良い
      */
     completionVariable() {
         // comp.kind = vscode.CompletionItemKind.Variable;
     }
-    //TODO:[1].storageとかならプロジェクト内のファイルパスを取得
     /**
      * ファイルの予測変換
      * InformationWorkSpaceに登録済みの素材Mapを取得すれば良い
      */
-    completionFile() {
-        //メモ
-        //storageパラメータとかでは、必要なフォルダのファイルだけを候補に出したい。例：bgのstorageならbgimageフォルダだけとか。
-        // storageパラメータのdescriptionに書いてある文字列によって分岐させる？背景、bgimageって単語があるならbgimage,音楽、BGM,があるならbgmフォルダとか。
-        // comp.kind = vscode.CompletionItemKind.File;
+    /**
+     *
+     * @param projectPath
+     * @param requireResourceType
+     * @param referencePath そのタグの参照するディレクトリのパス。例えば、bgタグならbgimageフォルダのパス
+     * @returns
+     */
+    async completionResource(projectPath, requireResourceType, referencePath) {
+        let completions = new Array();
+        this.infoWs.resourceFileMap.forEach((resourcesMap, key) => {
+            if (projectPath === key) {
+                resourcesMap.forEach((resource, key2) => {
+                    // if (resource.resourceType === requireResourceType) {
+                    if (requireResourceType.includes(resource.resourceType)) {
+                        let insertLabel = resource.filePath.replace(projectPath + this.infoWs.DATA_DIRECTORY + this.infoWs.pathDelimiter, "");
+                        insertLabel = insertLabel.replace(/\\/g, "/"); //パス区切り文字を/に統一
+                        let comp = new vscode.CompletionItem({
+                            label: resource.filePath.replace(projectPath + this.infoWs.DATA_DIRECTORY + this.infoWs.pathDelimiter, "").replace(/\\/g, "/"),
+                            description: resource.filePath.replace(projectPath + this.infoWs.pathDelimiter, ""),
+                            detail: ""
+                        });
+                        comp.kind = vscode.CompletionItemKind.File;
+                        comp.insertText = new vscode.SnippetString(path.relative(referencePath, resource.filePath).replace(/\\/g, "/")); //基準パスからの相対パス
+                        completions.push(comp);
+                    }
+                });
+            }
+        });
+        return completions;
     }
     //TODO:CompletionItemListを編集して必須のパラメータがわかるようにする
     /**
@@ -98,13 +139,16 @@ class TyranoCompletionItemProvider {
             if (selectedTag === tagName) {
                 for (const item2 in this.tyranoTagSnippets[item]["parameters"]) {
                     if (!(this.tyranoTagSnippets[item]["parameters"][item2]["name"] in parameters)) { //タグにないparameterのみインテリセンスに出す
-                        let comp = new vscode.CompletionItem(this.tyranoTagSnippets[item]["parameters"][item2]["name"]);
+                        const detailText = this.tyranoTagSnippets[item]["parameters"][item2]["required"] ? "（必須）" : "";
+                        const comp = new vscode.CompletionItem({
+                            label: this.tyranoTagSnippets[item]["parameters"][item2]["name"],
+                            description: "",
+                            detail: detailText
+                        });
                         comp.insertText = new vscode.SnippetString(this.tyranoTagSnippets[item]["parameters"][item2]["name"] + "=\"$0\" ");
                         comp.documentation = new vscode.MarkdownString(this.tyranoTagSnippets[item]["parameters"][item2]["description"]);
                         comp.kind = vscode.CompletionItemKind.Function;
-                        if (this.tyranoTagSnippets[item]["parameters"][item2]["required"]) {
-                            comp.label = this.tyranoTagSnippets[item]["parameters"][item2]["name"] + "(必須)";
-                        }
+                        comp.command = { command: 'editor.action.triggerSuggest', title: 'Re-trigger completions...' };
                         completions.push(comp);
                     }
                 }